Remove more from electron-browser/terminalService
